General clean up.

+++ src/GameState.cpp
@@ -3,7 +3,7 @@
 #include <ctime>
 
 GameState::GameState(sf::RenderWindow &renderWindow, sf::Font &textFont)
-	: State(renderWindow), turnManager(map), gui((sf::Vector2f) renderWindow.getSize()), mapBuilder(time(0), renderWindow, gui), font(textFont) {
+	: State(renderWindow), turnManager(map), gui((sf::Vector2f) renderWindow.getSize()), mapBuilder(static_cast<unsigned int>(time(0)), renderWindow, gui), font(textFont) {
 	map = mapBuilder.buildMap(sf::Vector2u(50, 50), sf::Vector2f(64, 64));
 	map->construct();
